requirements.txt
pygame
pillow
imageio
pyautogui